docs(Spec): Rewind now accepts a discrete tick number instead of a realtive argument.
For better control, rewind no longer accepts a relative number specifying the number of past states
to rewind. It now only accepts the parameter `t`, which must be a discrete tick number of a past
saved state.

BREAKING CHANGE: Any code passing a relative number to `rewind` must be switched to provide the
absolute tick number.