feat: Version 1.0.0-alpha.4
A complete rewrite of how inputs work.

BREAKING CHANGE: The StateMachine class now recognises the distinction between State and Input. A
second generic argument is need (`StateMachine<State, InputSource>`) that defined the layout of the
input source. The `@input` decorator must now decorate fields on the input source itself. The state
object passed to condition and actions is now a union of State & Input, but the return value from
actions can only be a partial State.