Version 1.0.0-alpha.4 (#3)
* feat: decorators are now chainable, added 2 new decorators

* feat: inhibitors via `@inhibitOn('methodName')`

* feat: exclusions via `@exceptWhen(condition)`

BREAKING CHANGE: @when is stricter and requires the state type as the first generic argument.

* chore: Switched the project to conventional-changelog.

Contributers must now commit any new changes using `npm run commit`.

* docs(Spec): Refined the spec.

Added decorators for @input and @priority to the spec. Fixed some typos and small errors in the
examples in the readme.

* docs(Spec): Fixed some typos.

* docs(Spec): Rewind now accepts a discrete tick number instead of a realtive argument.

For better control, rewind no longer accepts a relative number specifying the number of past states
to rewind. It now only accepts the parameter `t`, which must be a discrete tick number of a past
saved state.

BREAKING CHANGE: Any code passing a relative number to `rewind` must be switched to provide the absolute tick number.

* feat: Version 1.0.0-alpha.4

A complete rewrite of how inputs work.

BREAKING CHANGE: The `StateMachine` class now recognises the distinction between State and Input. A second generic argument is need (`StateMachine<State, InputSource>`) that defined the layout of the input source. The `@input` decorator must now decorate fields on the input source itself. The state object passed to condition and actions is now a union of State & Input, but the return value from actions can only be a partial State.